refactor feedback button into modal component and modify tests
