+++ frontend/src/components/AcmFeedbackModal.tsx
@@ -1,10 +1,14 @@
 /* Copyright Contributors to the Open Cluster Management project */
 
-import { FeedbackLocale, FeedbackModal, FeedbackModalProps } from '@patternfly/react-user-feedback'
+import { FeedbackLocale, FeedbackModal } from '@patternfly/react-user-feedback'
 import feedbackImage from './feedback_illo.png'
 import { useTranslation } from '../lib/acm-i18next'
+import { AcmButton } from '../ui-components'
+import { OutlinedCommentsIcon } from '@patternfly/react-icons'
+import { useState } from 'react'
+import { DOC_VERSION } from '../lib/doc-util'
 
-export const AcmFeedbackModal = (props: FeedbackModalProps) => {
+export const AcmFeedbackModal = () => {
   const { t } = useTranslation()
   const i18nStrings = {
     getSupport: t('Get support'),
@@ -16,15 +20,38 @@ export const AcmFeedbackModal = (props: FeedbackModalProps) => {
     tellAboutExperience: t('Tell us about your experience'),
   }
 
+  const [toggleOpen, setToggleOpen] = useState<boolean>(false)
+  const toggle = () => setToggleOpen(!toggleOpen)
+
   return (
-    <FeedbackModal
-      feedbackLocale={i18nStrings as unknown as FeedbackLocale}
-      email="test@redhat.com"
-      onShareFeedback={props.onShareFeedback}
-      onOpenSupportCase="https://access.redhat.com/support/cases/#/case/new/open-case?caseCreate=true"
-      isOpen={props.isOpen}
-      feedbackImg={feedbackImage}
-      onClose={props.onClose}
-    />
+    <>
+      <AcmButton
+        style={{
+          position: 'fixed',
+          transformOrigin: '50% -70%',
+          transform: 'rotate(270deg)',
+          bottom: '2em',
+          right: 0,
+          zIndex: 20000,
+          color: 'var(--pf-global--palette--white)',
+        }}
+        icon={<OutlinedCommentsIcon />}
+        iconPosition="left"
+        variant="danger"
+        id="feedback-trigger-button"
+        onClick={toggle}
+      >
+        {t('Feedback')}
+      </AcmButton>
+      <FeedbackModal
+        feedbackLocale={i18nStrings as unknown as FeedbackLocale}
+        email="test@redhat.com"
+        onShareFeedback={`https://console.redhat.com/self-managed-feedback-form?source=acm&version=${DOC_VERSION}`}
+        onOpenSupportCase="https://access.redhat.com/support/cases/#/case/new/open-case?caseCreate=true"
+        isOpen={toggleOpen}
+        feedbackImg={feedbackImage}
+        onClose={() => setToggleOpen(false)}
+      />
+    </>
   )
 }
